Debug.TracedInternal

Documentation

data Traced a

Traced values of some type.

traced :: Traceable a => a -> Traced a

Create a traced value.

named :: Traceable a => String -> a -> Traced a

Create a named traced value.

nameTraced :: Traceable a => String -> Traced a -> Traced a

Add a named to a traced value.

unknown :: Traceable a => String -> Traced a

Create a named thing with no value. Cannot be used where a real value is needed.

unTraced :: Traced a -> a

Extract the real value from a traced value.

tracedD :: Traceable a => Traced a -> TracedD

Extract the expression tree from a traced value.

data TraceLevel

Constructors

TLValue

TLExp

TLFullExp

class (Typeable a, Show a) => Traceable a where

Methods

trPPrint :: TraceLevel -> Int -> a -> Doc

data TracedD

Expression tree for a traced value.

Constructors

NoValue
unknown value

forall a . Name Bool Name TracedD
value with a name

forall a . Traceable a => Con a
constant

forall a . Traceable a => Apply a Name Fixity [TracedD]
application

forall a . Let [(Name, TracedD)] TracedD
(recovered) let expression

unTracedD :: Traceable a => TracedD -> Maybe (Traced a)

Convert an expression tree to a traced value, if the types are correct.

binOp :: (Traceable a, Traceable b, Traceable c) => (a -> b -> c) -> (String, Fixity) -> Traced a -> Traced b -> Traced c

unOp :: (Traceable a, Traceable b) => (a -> b) -> String -> Traced a -> Traced b

apply :: Traceable a => a -> Name -> Fixity -> [TracedD] -> Traced a

Create a traced value with an Apply expression tree.

liftT :: Liftable a b => Name -> Fixity -> a -> b

liftFun :: Liftable a b => Name -> a -> b

class Liftable a b | a -> b, b -> a where

Methods

liftT' :: Name -> Fixity -> [TracedD] -> a -> b

baseLiftT :: Traceable a => Name -> Fixity -> [TracedD] -> a -> Traced a

data Fixity

Fixity for identifier.

Constructors

InfixL Int

InfixR Int

Infix Int

Nonfix

showAsExp :: Traceable a => Traced a -> String

Show the expression tree of a traced value.

showAsExpFull :: Traceable a => Traced a -> String

Show the expression tree of a traced value, also show the value of each variable.

pPrintTraced :: Traceable a => TraceLevel -> Int -> Traced a -> Doc

reShare :: Traceable a => Traced a -> Traced a

simplify :: Traced a -> Traced a

Simplify an expression tree.