Debug.Traced

Description

The Traced module provides a simple way of tracing expression evaluation. A value of type Traced a has both a value of type a and an expression tree that describes how the value was computed.

There are instances for the Traced type for all numeric classes to make it simple to trace numeric expressions.

The expression tree associated with a traced value is exactly that: a tree. But evaluation of expressions in Haskell typically has sharing to avoid recomputation. This sharing can be recovered by the (impure) reShare function.

$examples

Documentation

data Traced a

Traced values of some type.

traced :: Traceable a => a -> Traced a

Create a traced value.

named :: Traceable a => String -> a -> Traced a

Create a named traced value.

nameTraced :: Traceable a => String -> Traced a -> Traced a

Add a named to a traced value.

unknown :: Traceable a => String -> Traced a

Create a named thing with no value. Cannot be used where a real value is needed.

unTraced :: Traced a -> a

Extract the real value from a traced value.

tracedD :: Traceable a => Traced a -> TracedD

Extract the expression tree from a traced value.

data TracedD

Expression tree for a traced value.

unTracedD :: Traceable a => TracedD -> Maybe (Traced a)

Convert an expression tree to a traced value, if the types are correct.

class (Typeable a, Show a) => Traceable a

liftT :: Liftable a b => Name -> Fixity -> a -> b

liftFun :: Liftable a b => Name -> a -> b

class Liftable a b | a -> b, b -> a

showAsExp :: Traceable a => Traced a -> String

Show the expression tree of a traced value.

showAsExpFull :: Traceable a => Traced a -> String

Show the expression tree of a traced value, also show the value of each variable.

reShare :: Traceable a => Traced a -> Traced a

simplify :: Traced a -> Traced a

Simplify an expression tree.

ifT :: Traceable a => Traced Bool -> Traced a -> Traced a -> Traced a

Traced version of if.

(%==), (%/=) :: (Traceable a, Eq a) => Traced a -> Traced a -> Traced Bool

Comparisons generating traced booleans.

(%<), (%>=), (%>), (%<=) :: (Traceable a, Ord a) => Traced a -> Traced a -> Traced Bool

(%&&) :: Traced Bool -> Traced Bool -> Traced Bool

(%||) :: Traced Bool -> Traced Bool -> Traced Bool

tnot :: Traced Bool -> Traced Bool

data TracedExp a

A wrapper for Traced to show it with full details.

tracedExp :: Traceable a => a -> TracedExp a

namedExp :: Traceable a => String -> a -> TracedExp a